+++ lib/honeybadger/plugins/rails.rb
@@ -31,11 +31,21 @@ module Honeybadger
 
       class ErrorSubscriber
         def self.report(exception, handled:, severity:, context: {}, source: nil)
-          return if source && ::Honeybadger.config[:'rails.subscriber_ignore_sources'].any? { |regex| regex.match?(source) }
+          # We only report unhandled errors (`Rails.error.handle`)
+          # Unhandled errors will be caught by our integrations (eg middleware), which have richer context
+          return unless handled
+
+          return if source_ignored?(source)
 
           tags = ["severity:#{severity}", "handled:#{handled}"]
           tags << "source:#{source}" if source
           Honeybadger.notify(exception, context: context, tags: tags)
+        end
+
+        def self.source_ignored?(source)
+          source && ::Honeybadger.config[:'rails.subscriber_ignore_sources'].any? do |ignored_source|
+            ignored_source.is_a?(Regexp) ? ignored_source.match?(source) : (ignored_source == source)
+          end
         end
       end
 
@@ -52,8 +62,7 @@ module Honeybadger
             ::ActionDispatch::ShowExceptions.prepend(ExceptionsCatcher)
           end
 
-          if defined?(::ActiveSupport::ErrorReporter)
-            # Rails 7
+          if defined?(::ActiveSupport::ErrorReporter) # Rails 7
             ::Rails.error.subscribe(ErrorSubscriber)
           end
         end